Drupal theme based on Library One-Pager with schema.org
First stab at separating out the relevant components of a Drupal theme,
plus several modules and view definitions, that replicates the look and
feel of the Library One-Pager site, but takes advantage of the CMS to
provide user-friendly widgets for editing the content and data. And most
importantly, automatically generates schema.org RDFa for that content.

Signed-off-by: Dan Scott <dan@coffeecode.net>